Some refactoring to separate MS-SRVS and MS-SCMR into their own files.
Added some more logging.

Added calculation of the applicationKey used in SMB 3.x and a public method to retrieve it.

Added some utility functions to facilitate handling of strings with NDR.

Added support to modify service configurations and specify a logon password for the service user.

Major overhaul on MS-SRVS and MS-SCMR encoding/decoding where the structs are made
simpler and more true to the MIDL specifications and the heavy lifting is moved to
the custom marshal/unmarshal functions.